Fix issue with sockets always being used.
When setting up this plugin for use on our own server we found that regardless of the fact that we can't even enable sockets it was using them. After digging through the code and your API I noticed the /socket endpoint returns both "socketAllowed" which is false and "socketEnabled" which is true.

While I would normally be fine with getting access to a feature for free, the plugin also decides to slow down the normal interval check when sockets are enabled meaning that purchases made when a user is offline take up to 35+ minutes to process once they join the server which just doesn't work. I'm only opening a PR for this change as I don't know internally how you would want to handle a solution to this other problem. Hope this helps.